+++ src/components/date-input/template.test.js
@@ -271,6 +271,17 @@ describe('Date input', () => {
       const $formGroup = $('.govuk-form-group')
       expect($formGroup.length).toBeTruthy()
     })
+
+    it('renders with a form group wrapper that has extra classes', () => {
+      const $ = render('date-input', {
+        formGroup: {
+          classes: 'extra-class'
+        }
+      })
+
+      const $formGroup = $('.govuk-form-group')
+      expect($formGroup.hasClass('extra-class')).toBeTruthy()
+    })
   })
 
   describe('when it includes a hint', () => {